Fixing ProductVariant test, which sometimes throws an error when cleaing up.

+++ ShopifySharp.Tests/ProductVariant_Tests.cs
@@ -116,6 +116,11 @@ namespace ShopifySharp.Tests
         {
             foreach (var obj in Created)
             {
+                if (! obj.Id.HasValue) 
+                {
+                    continue; 
+                }
+
                 try
                 {
                     await Service.DeleteAsync(ProductId, obj.Id.Value);
